simple_example_nodlet.cpp / simple_compressed_example_nodelet.cpp : support debug_view param

+++ src/nodelet/simple_compressed_example_nodelet.cpp
@@ -59,6 +59,7 @@ class ImageConverter
   ros::NodeHandle nh_;
   ros::Subscriber image_sub_;
   ros::Publisher image_pub_;
+  bool debug_view_;
 
 public:
   ImageConverter()
@@ -68,12 +69,18 @@ public:
             &ImageConverter::imageCb,this);
     image_pub_ = nh_.advertise<sensor_msgs::CompressedImage>("/image_converter/output_video/compressed", 1);
 
-    cv::namedWindow(OPENCV_WINDOW);
+    ros::NodeHandle pnh_("~");
+    pnh_.param("debug_view", debug_view_, false);
+    if( debug_view_) {
+      cv::namedWindow(OPENCV_WINDOW);
+    }
   }
 
   ~ImageConverter()
   {
-    cv::destroyWindow(OPENCV_WINDOW);
+    if( debug_view_) {
+      cv::destroyWindow(OPENCV_WINDOW);
+    }
   }
 
   void imageCb(const sensor_msgs::CompressedImageConstPtr& msg)
@@ -93,9 +100,11 @@ public:
     if (cv_ptr->image.rows > 110 && cv_ptr->image.cols > 110)
       cv::circle(cv_ptr->image, cv::Point(cv_ptr->image.cols/2, cv_ptr->image.rows/2), 100, CV_RGB(255,0,0));
 
-    // Update GUI Window
-    cv::imshow(OPENCV_WINDOW, cv_ptr->image);
-    cv::waitKey(3);
+    if( debug_view_) {
+      // Update GUI Window
+      cv::imshow(OPENCV_WINDOW, cv_ptr->image);
+      cv::waitKey(3);
+    }
 
     // Output modified video stream
     image_pub_.publish(cv_ptr->toCompressedImageMsg());
